clean up test/opt
moved tensor core tests together and removed some opts
